// Angular/Angular_profesjonalne_techniki_pracy/sk-planes/src/app/core/enums/role.enum.ts
export enum RoleLabel {
  Stewardess = 'Stewardess',
  SeniorCabinCrew = 'Senior Cabin Crew',
  Pilot = 'Pilot',
  CoPilot = 'Co Pilot',
  Mechanic = 'Mechanic',
}

export enum RoleValue {
  Stewardess = 'stewardess',
  SeniorCabinCrew = 'senior_cabin_crew',
  Pilot = 'pilot',
  CoPilot = 'co_pilot',
  Mechanic = 'mechanic',
}